[!!!][FEATURE] Added constant for media width
By default the media width is now read from the constant
styles.content.media.defaultVideoWidth (before it was
styles.content.imgtext.maxW).

Additionally a new constant was introduced to be able to use
a custom media width in mediaoembed:
plugin.tx_mediaoembed.defaultMaximumMediaWidth.

The constants are now in a category and have labels.