[FEATURE] Rewrite iframe src to SSL
If the current page is opened via a HTTPS request the src tag of an
iframe (only for video elements) is rewritten to HTTPS to make sure
that the browser does not block the request due to security restrictions.
Normally in a secure page only secure content is allowed, unsecure content
is blocked.