[FEATURE] Overridable aspect ratio for videos
The aspect ratio of videos can now be set / overwritten
in the Backend. This allows overriding invalid aspect
ratios from the providers an setting an aspect ratio
for providers that do not have width / height values in
their video responses.

If the provider does not return valid with / height values
the default fallback aspect ratio can be configured
in the TypoScript setting in 'aspectRatioFallback'.

The new tx_mediaoembed_aspect_ratio field is added to
the database and an AspectRatioEvaluation class is
added to make sure the field contains valid values.