[BUGFIX] Fix init of iframe URL parsing for inline elements
Instead of globally requireing the UrlParser JS in the
edit document view we introduce a new form field type
"tx_mediaoembed_urlinput".

This field is based on the TYPO3 core input field and adds
a wrapper div with a unique ID and initializes the JS
with the ID.

This makes sure that the URL parsing is initialized
even if the URL field is loaded via AJAX in an inline
record (like news content elements).